Delete ( false )

+++ front/src/components/project/ProjectCard.js
@@ -1,22 +1,37 @@
 import { Card, Button, Row, Col } from "react-bootstrap";
-import * as Api from "../../api"
+import * as Api from "../../api";
 
-function ProjectCard({ project,isEditable, setIsEditing,isDelete,setIsDelete }) {
-  const handleDelete = async (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-
-    // "project/수상 id" 엔드포인트로 Delete 요청함.
-    await Api.delete('projects', project.id);
-    setIsDelete(true)
-  };
+function ProjectCard({
+  project,
+  isEditable,
+  setIsEditing,
+  isDelete,
+  setIsDelete,
+}) {
+  async function handleDelete(e) {
+    if (window.confirm("정말 삭제하시겠습니까?")) {
+      e.preventDefault();
+      e.stopPropagation();
+      await Api.delete("projects", project.id);
+      setIsDelete(!isDelete);
+    }
+  }
   return (
     <Card.Body>
       <Row className="align-items-center">
         <Col>
-          <span>{project.title}</span>
+          <span className="primary">{project.title}</span>
           <br />
-          <span className="text-muted">{project.description}</span>
+          <span>{project.description}</span>
+          <br />
+          <span className="text-muted">
+            {project.from_date.toString().substr(0, 10)}
+          </span>
+          <br />
+          <span className="text-muted">
+            {project.to_date.toString().substr(0, 10)}
+          </span>
+          <br />
         </Col>
         {isEditable && (
           <Col xs lg="1">
@@ -28,10 +43,6 @@ function ProjectCard({ project,isEditable, setIsEditing,isDelete,setIsDelete }) 
             >
               편집
             </Button>
-          </Col>
-        )}
-        {!isDelete && (
-          <Col xs lg="1">
             <Button
               variant="outline-info"
               size="sm"
